requirements.txt
Flask==2.0.2
flask-cors==3.0.10
python-dotenv==0.19.2
marshmallow==3.14.1
google-generativeai==0.1.0
gunicorn==20.1.0
waitress==2.1.2